requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-JWT-Extended>=4.0.0
Flask-Bcrypt>=0.7.0
Flask-CORS>=3.0.0
Flask-Migrate>=4.0.0
Werkzeug>=2.0.0
Pillow>=8.0.0
email-validator>=1.1.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
